pyproject.toml
[tool.poetry]
name = "image-relayer"
version = "0.1.0"
description = "Discord bot to acquire images from various site URLs"
authors = ["Kenneth Lareau <elessar@numenor.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
PyYAML = "^5.4.1"
cloudscraper = "^1.2.58"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
